+++ channeldb/chanstate_assertions.go
@@ -1,0 +1,7 @@
+package channeldb
+
+import "github.com/lightningnetwork/lnd/chanstate"
+
+// Compile-time assertion that ChannelStateDB satisfies the channel-state store
+// contract while the KV implementation still lives in channeldb.
+var _ chanstate.Store[*OpenChannel] = (*ChannelStateDB)(nil)
