+++ channelnotifier/channelnotifier.go
@@ -18,7 +18,7 @@ type ChannelNotifier struct {
 
 	ntfnServer *subscribe.Server
 
-	chanDB chanstate.Store
+	chanDB chanstate.Store[*channeldb.OpenChannel]
 }
 
 // PendingOpenChannelEvent represents a new event where a new channel has
@@ -98,7 +98,7 @@ type FundingTimeoutEvent struct {
 // New creates a new channel notifier. The ChannelNotifier gets channel
 // events from peers and from the chain arbitrator, and dispatches them to
 // its clients.
-func New(chanDB chanstate.Store) *ChannelNotifier {
+func New(chanDB chanstate.Store[*channeldb.OpenChannel]) *ChannelNotifier {
 	return &ChannelNotifier{
 		ntfnServer: subscribe.NewServer(),
 		chanDB:     chanDB,
