+++ chanrestore.go
@@ -36,7 +36,7 @@ const (
 // need the secret key chain in order obtain the prior shachain root so we can
 // verify the DLP protocol as initiated by the remote node.
 type chanDBRestorer struct {
-	db chanstate.OpenChannelStore
+	db chanstate.OpenChannelStore[*channeldb.OpenChannel]
 
 	secretKeys keychain.SecretKeyRing
